Fix: ensure parent locator entries have enough space
