+++ packages/vhd-lib/src/Vhd/VhdDirectory.js
@@ -183,16 +183,7 @@ export class VhdDirectory extends VhdAbstract {
     return this._readChunk('parentLocator' + parentLocatorId)
   }
 
-  writeParentLocator(parentLocatorId, data) {
-    assert(parentLocatorId >= 0, 'parent Locator id must be a positive number')
-    assert(parentLocatorId < 8, 'parent Locator id  must be less than 8')
-    assert.notStrictEqual(this.header, undefined, `header must be read before it's used`)
-
-    const { platformDataSpace } = this.header.parentLocatorEntry[parentLocatorId]
-    if (!platformDataSpace) {
-      return
-    }
-    assert(data.length <= platformDataSpace)
+  _writeParentLocator(parentLocatorId, data) {
     return this._writeChunk('parentLocator' + parentLocatorId, data)
   }
 }
